+++ desenvolvimento/DonAct-Front/package.json
@@ -10,18 +10,11 @@
     "bootstrap-icons": "^1.9.1",
     "date-fns": "^2.29.1",
     "formik": "^2.2.9",
-    "icons": "^1.0.0",
-    "material": "^0.4.3",
-    "nodemon": "^2.0.19",
     "react": "^18.2.0",
     "react-bootstrap": "^2.5.0",
     "react-dom": "^18.2.0",
-    "react-elastic-carousel": "^0.11.5",
-    "react-hot-loader": "^4.13.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
-    "styled-components": "^5.3.5",
-    "web-vitals": "^2.1.4",
     "yup": "^0.32.11"
   },
   "scripts": {
