--- desenvolvimento/DonAct-Front/src/App.css
@@ -1,78 +1,0 @@
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-.App {
-  text-align: center;
-  overflow-x: hidden;
-  background-color: #FFF;
-  background-size:cover;
-  background-repeat: repeat;
-}
-.App-background{
-  background-color: #D8CBC7;
-  background-size:cover;
-  background-repeat: repeat;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  top: 0;
-}
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-.App-link {
-  color: #24572c;
-}
-
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.outer {
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  text-align: left;
-}
-
-.outer .form-control:focus {
-  border-color: #D8CBC7;
-  box-shadow: none;
-}
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.btn-primary{
-  background-color: #502223;
-  color: #FFF;
-  border-color: #502223;
-
-}
-  .btn-primary:hover{
-  background-color: #9D7B7B;
-  color: #FFF;
-  border-color: #502223;
-
-  }
-
