+++ desenvolvimento/DonAct-Front/src/components/Header/styles.css
@@ -1,0 +1,68 @@
+.header {
+	display: block;
+	width: 100vw;
+	margin: 0px;
+	font-family: Roboto;
+	padding: 0px;
+	top: 0;
+	position: relative;
+	height: 100%;
+}
+
+.donationimg{
+	width: 100%;
+    height: auto;
+	max-height: 32em;
+	z-index: 1;
+	position: relative;
+	top: 0;
+}
+
+.headerImg{
+	border-bottom: 3px;
+	border-color: #502223;
+	left: 0;
+	right: 0;
+	margin-bottom: 0;
+    border-bottom-style: solid;
+}
+
+.prospection{
+	font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+	align-items: flex-start;
+	align-self: flex-start;
+	align-content: flex-start;
+	position: relative;
+	top: 0;
+	margin-top: 4em;
+}
+
+.heartimg{
+	height: 4em;
+	margin-top: 2em;
+	margin-left: 6em;
+}
+
+.btn-singup{
+	display: flex;
+	margin-top: 2em;
+	margin-left: 14em;
+}
+
+.text-home{
+	display: inline-grid;
+	white-space: pre;
+	margin-left: 1.5em;
+	font-weight: bolder;
+	font-size: 2em;
+	color: #502223;
+	align-items: flex-start;
+}
+
+.bubblesimg{
+	position: absolute;
+	top: 0;
+	right: 0;
+	height: 30vw;
+	margin-top: -4em;
+}
