+++ desenvolvimento/DonAct-Front/src/components/Input/styles.css
@@ -1,0 +1,5 @@
+.label-login {
+  color: #502223;
+  font-weight: bold;
+  font-size: small;
+}
