+++ desenvolvimento/DonAct-Front/src/pages/ChoicePage/styles.css
@@ -1,0 +1,82 @@
+.inner {
+  width: auto;
+  margin: auto;
+  background: #ffffff;
+  box-shadow: 0px 14px 80px rgba(34, 35, 58, 0.2);
+  border-radius: 15px;
+  align-content: center;
+  transition: all 0.3s;
+}
+
+.inner-choice {
+  padding: 2em 4em 2em 3em;
+}
+
+.form {
+  align-content: center;
+}
+
+.form-title {
+  display: flex;
+  align-items: center;
+  white-space: nowrap;
+}
+
+.form-image {
+  max-width: 110px;
+  max-height: 200px;
+  display: inline-block;
+  margin-left: 40%;
+  position: relative;
+  text-align: center;
+  vertical-align: middle;
+}
+
+.form-begin-choice {
+  padding-top: 8%;
+  padding-bottom: 20%;
+}
+
+.h2-choice {
+  padding-top: 8% !important;
+  padding-left: 3% !important;
+  color: #886d6e;
+  font-family: "Cuprum", sans-serif;
+  display: block;
+}
+
+.btn-duo {
+  display: flex;
+  padding-left: 10%;
+  font-family: "Cuprum", sans-serif;
+  white-space: nowrap;
+  align-items: center;
+  padding-right: 10%;
+}
+
+.btn-tag {
+  align-content: center;
+  text-align: center;
+  display: inline-block;
+  padding-left: 3%;
+}
+
+.btn-choice {
+  margin-top: -10%;
+  text-align: center;
+  --bs-btn-padding-y: 2rem;
+  --bs-btn-padding-x: 1.5rem;
+}
+
+.vl {
+  border-left: 1px solid;
+  border-right: 1px solid;
+  background-color: #6a6a6a;
+  border-color: #6a6a6a;
+  border-radius: 7px;
+  height: 135px;
+  width: 3px;
+  padding: 0;
+  margin-left: 0;
+  position: static;
+}
