--- desenvolvimento/DonAct-Front/src/pages/Home.js
@@ -1,15 +1,0 @@
-import React from "react";
-import Header from "../components/Header";
-import Footer from "../components/Footer";
-import Navbar from "../components/Navbar";
-
-const Home = (props) => {
-    return (
-        <div id="body-container">
-            <Navbar/>
-            <Header/>
-	    </div>
-    );
-
-};
-    export default Home;
