+++ desenvolvimento/DonAct-Front/src/pages/HomePage/index.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+import Header from "../../components/Header";
+import Navbar from "../../components/Navbar";
+
+const Home = () => {
+  return (
+    <div id="body-container">
+      <Navbar />
+      <Header />
+    </div>
+  );
+};
+export default Home;
