+++ desenvolvimento/DonAct-Front/src/pages/HomePage/styles.css
@@ -1,0 +1,3 @@
+body {
+	overflow-x: hidden;
+}
