+++ desenvolvimento/DonAct-Front/src/pages/LoginPage/styles.css
@@ -1,0 +1,94 @@
+.form {
+  align-content: center;
+}
+
+.form-btn {
+  align-items: center;
+  padding-left: 35%;
+  padding-top: 8%;
+}
+.form.group {
+  display: flex;
+}
+
+.form-begin {
+  padding-top: 20%;
+  padding-bottom: 30%;
+}
+
+.form-image {
+  max-width: 110px;
+  max-height: 99px;
+  display: inline-block;
+  margin-left: 75%;
+  position: relative;
+  text-align: center;
+  vertical-align: middle;
+}
+
+.form-title {
+  display: flex;
+  align-items: center;
+  white-space: nowrap;
+}
+
+.form-btn {
+  align-items: center;
+  padding-left: 30%;
+  padding-top: 8%;
+}
+
+.register-text {
+  color: #212121;
+  padding-top: 10%;
+  font-size: smaller;
+}
+
+.register-text a {
+  color: #653c3d;
+}
+
+.register-text a:hover {
+  color: #9c8081;
+}
+
+.inner {
+  width: auto;
+  margin: auto;
+  background: #ffffff;
+  box-shadow: 0px 14px 80px rgba(34, 35, 58, 0.2);
+  border-radius: 15px;
+  align-content: center;
+  transition: all 0.3s;
+}
+
+.inner-login {
+  padding: 2em 2.5em 2em 2.5em;
+}
+
+.forgot-password,
+.forgot-password a {
+  text-align: right;
+  font-size: 13px;
+  padding-top: 10px;
+  color: #7f7d7d;
+  margin: 0;
+}
+
+.forgot-password a {
+  color: #502223;
+}
+
+.forgot-password a:hover {
+  color: #6d5252;
+}
+
+.outer {
+  background-color: #d8cbc7;
+}
+
+.label-login {
+  color: #502223;
+  font-weight: bold;
+  font-size: small;
+}
