--- desenvolvimento/DonAct-Front/src/styles/footer.css
@@ -1,73 +1,0 @@
-
-#Footer {
-    width: auto;
-    height: 3em;
-    margin-left: 0.2em;
-    margin-top: 2em;
-    font-size: 1.5em;
-    color: #653C3D;
-    /* background-color: cornsilk; */
-    font-weight: bold;
-    
-}
-
-#Contacts{
-    width: 12em;
-    height: 5em;
-    /* background-color:rgb(163, 163, 211); */
-    font-size: 0.8em;
-    font-weight: normal;
-    margin-top: 0.5em;
-}
-
-#Contacts_direita{
-    width: 50%;
-    height: 100%;
-    float: right;
-    padding-top: 10%;
-    /* background-color:rgb(144, 224, 177); */
-    
-}
-
-#Contacts_esquerda{
-    width: 50%;
-    height: 100%;
-    float: left;
-    padding-top: 10%;
-    border-right-style: solid;
-    border-width: 0.01em;
-    /* background-color:rgb(224, 189, 144); */
-    
-}
-
-#Buttons_Footer{
-    width: 1em;
-    height: 1.5em;
-    float: right;
-    padding-top: 1em;
-    margin-right:0em;
-
-}
-
-#Buttons_Footer button{
-    width: 3em;
-    height: 3em;
-    border-radius: 50%;
-    margin-left: 1em;
-    border: none;
-    background: none;
-
-    
-}
-
-#Buttons_Footer img{
-    width: 3.7em;
-}
-
-.img{
-    max-width: 115px;
-    max-height: 50px;
-    margin-top: 0.3em;
-    margin-bottom: 0.3em;
-    margin-left: 2em;
-}
