--- desenvolvimento/DonAct-Front/src/styles/home.css
@@ -1,7 +1,0 @@
-body {
-	overflow-x: hidden;
-}
-.body-container{
-	background-color: #FFF;
-	height: 100%;
-}
