--- desenvolvimento/DonAct-Front/src/styles/navbar.css
@@ -1,47 +1,0 @@
-.navbar {
-
-    width: 100%;
-    display: flex;
-    position: fixed;
-    background-color: #502223;
-    top: 0;
-    right: 0;
-    left: 0;
-    z-index: 99;
-}
-
-.logonav{
-    max-width: 115px;
-    max-height: 50px;
-    margin-top: 0.1em;
-    margin-bottom: 0.1em;
-    margin-left: 2em;
-}
-.login{
-    margin-top: 0.3em;
-    margin-bottom: 0.3em;
-    margin-right: 2em;
-}
-.loginButton{
-    display: flex;
-    margin-top: 0.1em;
-    margin-bottom: 0.1em;
-    margin-left: 2em;
-    background-color: #653C3D;
-    border-width: 2px ;
-    border-color: #D8CBC7;
-}
-.loginButton:hover{
-    border-color: #D8CBC7;
-    background-color: #502223;
-}
- span {
-	font-size: 3.5em;
-	cursor: pointer;
-	margin-left: 5%;
-    color:#D8CBC7 ;
-}
-
-label{
-    color:#D8CBC7 ;
-}
